Make constants private to generate less documentation
crates.io wouldn't let me publish the crate as it was above the size
limit.
By making the constants private, I can expose a much smaller API and
reduce the documentation bloat significantly.

enums would be a better way to get to these characters, so it isn't so
bad to loose this little bit of API flexibility.